+++ src/Libraries/AddrArrayLib.sol
@@ -15,7 +15,7 @@ library AddrArrayLib {
     using AddrArrayLib for Addresses;
 
     struct Addresses {
-      address[]  _items;
+        address[] _items;
     }
 
     /**
@@ -25,9 +25,9 @@ library AddrArrayLib {
      * @param element the element to add in the array
      */
     function pushAddress(Addresses storage self, address element) internal {
-      if (!exists(self, element)) {
-        self._items.push(element);
-      }
+        if (!exists(self, element)) {
+            self._items.push(element);
+        }
     }
 
     /**
@@ -38,7 +38,7 @@ library AddrArrayLib {
      * @param element the element to remove from the array
      */
     function removeAddress(Addresses storage self, address element) internal returns (bool) {
-        for (uint i = 0; i < self.size(); i++) {
+        for (uint256 i = 0; i < self.size(); i++) {
             if (self._items[i] == element) {
                 self._items[i] = self._items[self.size() - 1];
                 self._items.pop();
@@ -64,7 +64,7 @@ library AddrArrayLib {
      * @param self Storage array containing address type variables
      */
     function size(Addresses storage self) internal view returns (uint256) {
-      return self._items.length;
+        return self._items.length;
     }
 
     /**
@@ -73,7 +73,7 @@ library AddrArrayLib {
      * @param element the element to check if it exists in the array
      */
     function exists(Addresses storage self, address element) internal view returns (bool) {
-        for (uint i = 0; i < self.size(); i++) {
+        for (uint256 i = 0; i < self.size(); i++) {
             if (self._items[i] == element) {
                 return true;
             }
@@ -85,8 +85,7 @@ library AddrArrayLib {
      * @notice get the array
      * @param self Storage array containing address type variables
      */
-    function getAllAddresses(Addresses storage self) internal view returns(address[] memory) {
+    function getAllAddresses(Addresses storage self) internal view returns (address[] memory) {
         return self._items;
     }
-
 }
